res updated some code to use new-style endian conversions. res deprecated old-style endian conversions. res added support for optimized byte swapping, when supported by   the compiler. res removed support for MD2 conversions from md22spr (since MD2s   can now be loaded via the spr3md2 plugin). Also renamed md22spr to   mdl2spr. res changed wxtest to use RequestPlugins() to load the WX canvas.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23457 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/import/md22spr/m2s_base.h
@@ -24,7 +24,7 @@
 #include "cstypes.h"
 #include <stdio.h>
 
-typedef float scalar_t;       // Scalar value,
+typedef uint32 scalar_t;       // Scalar value,
 
 struct vec3_t                 // Vector or Position
 {
